+++ src/drone.c
@@ -8,6 +8,7 @@
 #include <fcntl.h>
 #include <sys/file.h>
 #include <sys/wait.h>
+#include <time.h>
 #include <signal.h>
 #include "logger.h"
 #include "parameter_file.h"
@@ -75,21 +76,24 @@ int main(int argc, char *argv[])
     log_config(FILENAME_LOG, LOG_DEBUG);
 
     WritePid();
-    kill(watchdogPid, SIG_WRITTEN);
+    if (watchdogPid != -1) kill(watchdogPid, SIG_WRITTEN);
     
     //Protocol messages
     Msg_int msg_int_in;
     Msg_float msg_float_in, msg_float_out;
 
-    //Signal from watchdog
-    struct sigaction sa_ping;
-    sa_ping.sa_handler = ping_handler;
-    sa_ping.sa_flags = SA_RESTART;
-    sigemptyset(&sa_ping.sa_mask);
-    
-    if (sigaction(SIG_PING, &sa_ping, NULL) == -1) {
-        perror("Error in ping_handler");
-        exit(EXIT_FAILURE);
+    if (watchdogPid != -1)
+    {
+        //Signal from watchdog
+        struct sigaction sa_ping;
+        sa_ping.sa_handler = ping_handler;
+        sa_ping.sa_flags = SA_RESTART;
+        sigemptyset(&sa_ping.sa_mask);
+        
+        if (sigaction(SIG_PING, &sa_ping, NULL) == -1) {
+            perror("Error in ping_handler");
+            exit(EXIT_FAILURE);
+        }
     }
 
 
@@ -100,22 +104,29 @@ int main(int argc, char *argv[])
     drone drn;
     
     bool exiting = false;
-
-    //Initializzation drone proprieties
-    drn.x=init_x;
-    drn.y=init_y;
-    drn.Fx=0;
-    drn.Fy=0;
-    drn.x_1=init_x;
-    drn.y_1=init_y;
-    drn.x_2=init_x;
-    drn.y_2=init_y;
     
-	
     //Read the initial dimension of the window
 	read(fd_r_drone, &msg_int_in, sizeof(msg_int_in));
     size.height = msg_int_in.a;
     size.width = msg_int_in.b;
+
+    if(msg_int_in.type == 's') srand(time(NULL) + 10);
+    else srand(time(NULL) + 253);
+
+    int x_rand = rand() % (size.width - 2) + 1; 
+    int y_rand = rand() % (size.height - 2) + 1;
+
+    //Initializzation drone proprieties
+    drn.x=x_rand;
+    drn.y=y_rand;
+    drn.Fx=0;
+    drn.Fy=0;
+    drn.x_1=x_rand;
+    drn.y_1=y_rand;
+    drn.x_2=x_rand;
+    drn.y_2=y_rand;
+    
+	
 
 
     while(1)
