+++ src/obstacles.c
@@ -84,6 +84,8 @@ void ObstacleRepulsion(bool array[][MaxWidth], float x, float y, float *Fx, floa
 
     if (min_c < 0) min_c = 0;
     if (min_r < 0) min_r = 0;
+
+    log_debug("Calculating obstacle repulsion for position: %f, %f", x, y);
     
     for (int r = min_r; r <= max_r; r++)
     {
@@ -104,7 +106,7 @@ void ObstacleRepulsion(bool array[][MaxWidth], float x, float y, float *Fx, floa
                     //Latombe's formula
                     //F = eta * (1/d - 1/rho) * (1/d^2)
                     float term1 = (1.0f / d) - (1.0f / rho);
-                    float F = eta * term1 * (1.0f / (d * d));
+                    float F = eta_obstacle * term1 * (1.0f / (d * d));
 
                     //(dx/d) is the cosine
                     //(dy/d) is the sine
@@ -132,7 +134,7 @@ void WallRepulsion(float x, float y, float *Fx, float *Fy, int height, int width
 
         //Calculate relative distance 
         float term1 = (1.0f / d) - (1.0f / rho);
-        float F = eta * term1 * (1.0f / (d * d));
+        float F = eta_obstacle * term1 * (1.0f / (d * d));
 
         //if (F > MaxRepulsive) F = MaxRepulsive;
         log_debug("The d from wall LEFT is %f and the force is %f", d, F);
@@ -149,7 +151,7 @@ void WallRepulsion(float x, float y, float *Fx, float *Fy, int height, int width
 
         //Calculate relative distance 
         float term1 = (1.0f / d) - (1.0f / rho);
-        float F = eta * term1 * (1.0f / (d * d));
+        float F = eta_obstacle * term1 * (1.0f / (d * d));
 
         //if (F > MaxRepulsive) F = MaxRepulsive;
         log_debug("The d from wall RIGHT is %f and the force is %f", d, F);
@@ -166,7 +168,7 @@ void WallRepulsion(float x, float y, float *Fx, float *Fy, int height, int width
 
         //Calculate relative distance 
         float term1 = (1.0f / d) - (1.0f / rho);
-        float F = eta * term1 * (1.0f / (d * d));
+        float F = eta_obstacle * term1 * (1.0f / (d * d));
 
         //if (F > MaxRepulsive) F = MaxRepulsive;
         log_debug("The d from wall UP is %f and the force is %f", d, F);
@@ -183,7 +185,7 @@ void WallRepulsion(float x, float y, float *Fx, float *Fy, int height, int width
 
         //Calculate relative distance 
         float term1 = (1.0f / d) - (1.0f / rho);
-        float F = eta * term1 * (1.0f / (d * d));
+        float F = eta_obstacle * term1 * (1.0f / (d * d));
 
         //if (F > MaxRepulsive) F = MaxRepulsive;
         log_debug("The dfrom wall BOTTOM is %f and the force is %f", d, F);
@@ -234,6 +236,11 @@ void Positioning(bool array[][MaxWidth], int height, int width)
 
 }
 
+void PositioningEnemy(bool array[][MaxWidth], int x, int y)
+{
+    array[y][x] = true;
+}
+
 int main(int argc, char *argv[])
 {
     int fd_r_obstacle, fd_w_obstacle;
@@ -241,7 +248,7 @@ int main(int argc, char *argv[])
 
     log_config(FILENAME_LOG, LOG_DEBUG);
     WritePid();
-    kill(watchdogPid, SIG_WRITTEN);
+    if(watchdogPid != -1) kill(watchdogPid, SIG_WRITTEN);
     
     bool obstacle[MaxHeight][MaxWidth];     //Max dimension of the screen
     
@@ -250,16 +257,20 @@ int main(int argc, char *argv[])
     Msg_int msg_int_out;
     Msg_float msg_float_in, msg_float_out;
     int h_Win, w_Win;
+    int enemy_x, enemy_y;
 
-    //Signal from watchdog
-    struct sigaction sa_ping;
-    sa_ping.sa_handler = ping_handler;
-    sa_ping.sa_flags = SA_RESTART;
-    sigemptyset(&sa_ping.sa_mask);
-    
-    if (sigaction(SIG_PING, &sa_ping, NULL) == -1) {
-        perror("Error in ping_handler");
-        exit(EXIT_FAILURE);
+    if(watchdogPid != -1)
+    {
+        //Signal from watchdog
+        struct sigaction sa_ping;
+        sa_ping.sa_handler = ping_handler;
+        sa_ping.sa_flags = SA_RESTART;
+        sigemptyset(&sa_ping.sa_mask);
+        
+        if (sigaction(SIG_PING, &sa_ping, NULL) == -1) {
+            perror("Error in ping_handler");
+            exit(EXIT_FAILURE);
+        }
     }
 
     float Fx, Fy;
@@ -278,7 +289,10 @@ int main(int argc, char *argv[])
                 w_Win = (int)msg_float_in.b;
                 log_debug("Window dimension: %d, %d", h_Win, w_Win);
                 ClearArray(obstacle);
-                Positioning(obstacle, h_Win, w_Win);
+                if(watchdogPid != -1)       //If single player
+                {
+                    Positioning(obstacle, h_Win, w_Win);
+                }
                 write(fd_w_obstacle, obstacle, sizeof(obstacle));
                 break;
             case 'f':       //The BB wants to know the forces that obstacles apply to the drone
@@ -291,9 +305,14 @@ int main(int argc, char *argv[])
                 Set_msg(msg_float_out, 'f', Fx, Fy);
                 write(fd_w_obstacle, &msg_float_out, sizeof(msg_float_out));
                 break;
+            case 'm':       //Multiplayer case: the BB tell us where is the drone of the other player
+                enemy_x = (int)msg_float_in.a;
+                enemy_y = (int)msg_float_in.b;
+                ClearArray(obstacle);
+                PositioningEnemy(obstacle, enemy_x, enemy_y);
+                break;
             default:
                 log_error("Error: wrong format of the message recived");
-                perror("Format not correct");
                 break;
         }
